export ToResult trait use it instead of checking if return_type starts with Result
