+++ macros/src/lib.rs
@@ -6,7 +6,7 @@ mod var_type;
 use attributes::Attributes;
 use export_function::{
     args_from_function_params, args_import_from_function_params, function_params_from_itemfn,
-    get_body, get_return_type, params_from_function_params,
+    get_return_type, params_from_function_params,
 };
 use proc_macro::TokenStream;
 use quote::quote;
@@ -24,8 +24,7 @@ pub fn export_function(attr: TokenStream, function: TokenStream) -> TokenStream 
     let args = args_from_function_params(&params);
     let args_importing = args_import_from_function_params(&params);
     let params = params_from_function_params(&params);
-    let (return_type, is_result) = get_return_type(&function, attr.return_type);
-    let body = get_body(is_result, &ident, &args);
+    let return_type = get_return_type(&function, attr.return_type);
     quote!(
         #function
         {
@@ -35,7 +34,9 @@ pub fn export_function(attr: TokenStream, function: TokenStream) -> TokenStream 
                     simplesl::function::Params(std::sync::Arc::new([#params])),
                     |interpreter| {
                         #args_importing
-                        #body
+                        simplesl::ToResult::<_, simplesl::errors::ExecError>::to_result(
+                            #ident(#args)
+                        ).map(|value| value.into())
                     },
                     #return_type,
                 ).into(),
